Update Makefile (ISO/IEC 9899:1999)
Some distros like RedHat or Centos do not implement by default the ISO c99 mode in the gcc compiler. Therefore when compiling the proof of concept it shows errors on all lines where the variable was defined inside a loop. Adding this parameter will help it compile on all GNU/Linux distributions.